requirements.txt
alpaca==1.0.0
alpaca-py==0.38.0
alpaca-trade-api==3.2.0
matplotlib==3.10.0
numpy==2.2.2
openai==1.61.1
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
scikit-learn==1.6.1
scipy==1.15.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
